Add files via upload
In these .csv files, employee_preprocess_200638P.csv is the preprocessed file of employees.csv.  
In the 200638P.py file, the employee data set was preprocessed and the null values were imputed using a decision tree regressor and classifier.
The 20063P.pdf file consists of the data description, how the data preprocessing was done, and the insights that were gained.